backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.5
python-multipart==0.0.6
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
alembic==1.13.1
aiosqlite==0.19.0
python-dotenv==1.0.0
httpx==0.25.2
sse-starlette==1.6.5
argon2-cffi==21.3.0